# backend/src/main/java/ru/hackathon/airballoon/admin/dto/LoginRequest.java
package ru.hackathon.airballoon.admin.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LoginRequest(
        @NotBlank @Size(max = 100) String username,
        @NotBlank @Size(max = 200) String password) {
}